* commit '5149a68f85cfb17b1c85853dd4bb97c0a84624bf':
  replace size_t with int to have safe integer comparison
  revert setting of index 11 with electron PDG code

+++ analysis/variables/src/PIDVariables.cc
@@ -333,10 +333,10 @@ namespace Belle2 {
         B2ERROR("The variable binaryPID_SVD needs exactly two arguments: the PDG codes of two hypotheses.");
         return std::numeric_limits<float>::quiet_NaN();
       }
-      size_t pdgCodeHyp = std::abs(int(std::lround(arguments[0])));
-      size_t pdgCodeTest = std::abs(int(std::lround(arguments[1])));
-      std::vector<size_t> pdgIds {pdgCodeHyp, pdgCodeTest};
-      if (std::any_of(pdgIds.begin(), pdgIds.end(), [](size_t p) {return (p == Const::electron.getPDGCode() || p == Const::muon.getPDGCode() || p == Const::deuteron.getPDGCode());})) {
+      int pdgCodeHyp = std::abs(int(std::lround(arguments[0])));
+      int pdgCodeTest = std::abs(int(std::lround(arguments[1])));
+      std::vector<int> pdgIds {pdgCodeHyp, pdgCodeTest};
+      if (std::any_of(pdgIds.begin(), pdgIds.end(), [](int p) {return (p == Const::electron.getPDGCode() || p == Const::muon.getPDGCode() || p == Const::deuteron.getPDGCode());})) {
         B2ERROR("The variable binaryPID_SVD is not defined for particle hypotheses {11, 13, 1000010020}.");
         return std::numeric_limits<float>::quiet_NaN();
       }
